requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
mcp>=0.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
uvicorn>=0.27.0
fastapi>=0.109.0
httpx>=0.26.0
openai>=1.0.0
